requirements.txt
Flask==3.0.0
gunicorn==21.2.0
Pillow==11.3.0
pillow-heif==1.1.1
python-dotenv==1.0.0
Werkzeug==3.0.0
